Make serialization more flexible and add Agent sample
Implements new serialization approach where a builder maintains a
reference to the serializer for its block. In order to enable
abstracting over the type of the environment, a block's environment
remains in serialized form until the type of its environment is
known. This requires a read-writer for block data with serialized
environments.

Also adds a sample that shows how to implement Agents which receive
serialized blocks. The received blocks are deserialized and used to
update the state maintained by the Agent.